fix(db): move ListingRow and getListingById to @ld/db (#65)
- What: extracted duplicated ListingRow interface and getListingById()
  from both listing-processor and upload-api into the shared @ld/db package
- Why: schema changes required updating the type in two places; query
  lived in two files with identical implementations
- How verified: pnpm build succeeds; both packages re-export from @ld/db
  so downstream imports are unchanged
