+++ lib/db/src/index.ts
@@ -7,6 +7,36 @@ import type Database from 'better-sqlite3';
  * Apply the shared DDL (listings + upload_log tables) to the given database.
  * Each service may apply additional service-specific tables after calling this.
  */
+// ── Types ───────────────────────────────────────────────────────────
+
+export interface ListingRow {
+  id: number;
+  odoo_product_id: number;
+  odoo_product_name: string | null;
+  status: string;
+  listing_data: string;
+  title: string | null;
+  price: number | null;
+  ebay_item_id: string | null;
+  ebay_url: string | null;
+  error_message: string | null;
+  notes: string | null;
+  created_at: string;
+  approved_at: string | null;
+  uploaded_at: string | null;
+}
+
+// ── Shared Queries ──────────────────────────────────────────────────
+
+export function getListingById(
+  db: Database.Database,
+  id: number,
+): ListingRow | undefined {
+  return db.prepare('SELECT * FROM listings WHERE id = ?').get(id) as ListingRow | undefined;
+}
+
+// ── Schema ──────────────────────────────────────────────────────────
+
 export function applySchema(db: Database.Database): void {
   db.exec(`
     CREATE TABLE IF NOT EXISTS listings (
