+++ packages/listing-processor/src/db.ts
@@ -4,7 +4,8 @@
 import Database from 'better-sqlite3';
 import { join } from 'node:path';
 import { homedir } from 'node:os';
-import { applySchema } from '@ld/db';
+import { applySchema, type ListingRow } from '@ld/db';
+export { type ListingRow, getListingById } from '@ld/db';
 
 const DB_PATH = process.env.DB_PATH ?? join(homedir(), 'ebay-listings.db');
 
@@ -26,25 +27,6 @@ export function getDb(): Database.Database {
   return _db;
 }
 
-// ── Types ───────────────────────────────────────────────────────────
-
-export interface ListingRow {
-  id: number;
-  odoo_product_id: number;
-  odoo_product_name: string | null;
-  status: string;
-  listing_data: string;
-  title: string | null;
-  price: number | null;
-  ebay_item_id: string | null;
-  ebay_url: string | null;
-  error_message: string | null;
-  notes: string | null;
-  created_at: string;
-  approved_at: string | null;
-  uploaded_at: string | null;
-}
-
 // ── Queries ─────────────────────────────────────────────────────────
 
 export function getListingByProductId(
@@ -54,13 +36,6 @@ export function getListingByProductId(
   return db.prepare(
     'SELECT * FROM listings WHERE odoo_product_id = ?'
   ).get(productId) as ListingRow | undefined;
-}
-
-export function getListingById(
-  db: Database.Database,
-  id: number,
-): ListingRow | undefined {
-  return db.prepare('SELECT * FROM listings WHERE id = ?').get(id) as ListingRow | undefined;
 }
 
 export function getListingsByStatus(
